Added Docker Maven plugin
To use the plugin first run

mvn package
mvn docker:build

By default the Docker Maven plugin will look for the Dockerfile in src/main/Docker.

The plugin runs from target/docker/kbe-rest/build and the artifact is copied to the /maven sub-folder that is why we needed to change from "ADD target" to "ADD maven".